Add 2 pages, images, and files
Landing page links to about and contact pages and theres a large photo on this home page. About page has a short blurb and a link to a pdf of my current resume. The contact page has my email and github. There are more photos in the images folder that need to be placed, probably in the about page.