Make several display changes to nav bar and site
Nav bar is now float right. about page text is now center justified by nesting the p in a div tag. The div tag was center aligned and the daughter p tag was set to justified. The links on the contact page were also made live to the github page and mailto the email address.